requirements/base.txt
psycopg2==2.6.1
Pillow==2.9.0
Django==1.11.1
six==1.9.0

BeautifulSoup4==4.4.0

django-dotenv==1.3.0
dealer==2.0.5
djangorestframework==3.6.2
# Markdown support for the browsable API.
Markdown==2.6.8
# Filtering support
django-filter==1.0.2
